Merge Roomuery -> develop

+++ Query/Types/RoomMember.cs
@@ -6,8 +6,8 @@ namespace SharpChatwork.Query.Types
 {
     public class RoomMember
     {
-        public List<int> admin { get; set; } = new List<int>();
-        public List<int> member { get; set; } = new List<int>();
-        public List<int> @readonly { get; set; } = new List<int>();
+        public List<long> admin { get; set; } = new List<long>();
+        public List<long> member { get; set; } = new List<long>();
+        public List<long> @readonly { get; set; } = new List<long>();
     }
 }
